requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-community==0.0.10
langchain-core==0.1.23
faiss-cpu==1.7.4
pypdf==3.17.4
python-docx==1.1.0
python-dotenv==1.0.0
sentence-transformers>=2.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

